center item for login and register page
