+++ SHOPHVT/Content/style.css
@@ -9024,7 +9024,7 @@ h1, h2, h3, h4, h5,
   padding-left: 15px;
   padding-right: 15px;
   background: transparent !important;
-  color: rgba(0, 0, 0, 0.4) !important;
+  color: rgba(0, 0, 0, 1) !important;
   font-size: 14px;
   border-radius: 0px;
   -webkit-box-shadow: none !important;
