+++ SHOPHVT/Controllers/ProductController.cs
@@ -47,5 +47,19 @@ namespace SHOPHVT.Views
             ViewBag.Categories = categories;
             return View(products);
         }
+
+        public ActionResult Details(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(System.Net.HttpStatusCode.BadRequest);
+            }
+            SanPham sanPham = db.SanPhams.Find(id);
+            if (sanPham == null)
+            {
+                return HttpNotFound();
+            }
+            return View(sanPham);
+        }
     }
 }
