+++ SHOPHVT/Views/Cart/Paycheck.cshtml
@@ -38,7 +38,7 @@
                         <div class="col-md-6">
                             <div class="form-group">
                                 <label for="phone">Phone</label>
-                                <input type="text" class="form-control" placeholder="">
+                                <input type="text" class="form-control" placeholder="@ViewBag.CustomerPhoneNumber">
                             </div>
                         </div>
                         <div class="col-md-6">
