+++ SHOPHVT/Views/Customer/Update.cshtml
@@ -19,7 +19,7 @@
         <div class="container">
             <div class="row justify-content-center">
                 <div class="col-xl-7 ftco-animate">
-                    <form method="post" class="">
+                    <form method="post" class="billing-form">
                         <h3 class="mb-4 billing-heading">Cập nhật thông tin tài khoản</h3>
                         <div class="row align-items-end">
                             <div class="col-md-12">
@@ -61,6 +61,7 @@
                                     <label for="Gender">Giới tính</label>
                                     @*<input name="Gender" type="text" class="form-control" placeholder="" value="@Model.Gender">*@
                                     <div class="select-wrap">
+                                        <div class="icon"><span class="ion-ios-arrow-down"></span></div>
                                         <select name="Gender" id="Gender" class="form-control">
                                             <option selected="selected">Nam</option>
                                             <option>Nữ</option>
