+++ SHOPHVT/Views/Customer/UpdatePassword.cshtml
@@ -19,7 +19,7 @@
         <div class="container">
             <div class="row justify-content-center">
                 <div class="col-xl-7 ftco-animate">
-                    <form method="post" class="">
+                    <form method="post" class="billing-form">
                         <h3 class="mb-4 billing-heading">Cập nhật mật khẩu</h3>
                         @Html.ValidationSummary()
                         <div class="row align-items-end">
