+++ SHOPHVT/Views/KhachHang/Create.cshtml
@@ -40,6 +40,15 @@
                     @Html.ValidationMessageFor(model => model.gioiTinh, "", new { @class = "text-danger" })
                 </div>
             </div>
+
+            <div class="form-group">
+                <div class="control-label col-md-2">SDT</div>
+                <div class="col-md-10">
+                    @Html.EditorFor(model => model.soDienThoai, new { htmlAttributes = new { @class = "form-control" } })
+                    @Html.ValidationMessageFor(model => model.soDienThoai, "", new { @class = "text-danger" })
+                </div>
+            </div>
+
             <div class="form-group">
                 <div class="control-label col-md-2">Email</div>
                 <div class="col-md-10">
