+++ SHOPHVT/Views/KhachHang/Delete.cshtml
@@ -37,6 +37,14 @@
             </dd>
 
             <dt>
+                @Html.DisplayNameFor(model => model.soDienThoai)
+            </dt>
+
+            <dd>
+                @Html.DisplayFor(model => model.soDienThoai)
+            </dd>
+
+            <dt>
                 @Html.DisplayNameFor(model => model.email)
             </dt>
