+++ SHOPHVT/Views/KhachHang/Details.cshtml
@@ -36,6 +36,14 @@
             </dd>
 
             <dt>
+                @Html.DisplayNameFor(model => model.soDienThoai)
+            </dt>
+
+            <dd>
+                @Html.DisplayFor(model => model.soDienThoai)
+            </dd>
+
+            <dt>
                 @Html.DisplayNameFor(model => model.email)
             </dt>
