+++ SHOPHVT/Views/KhachHang/Edit.cshtml
@@ -44,6 +44,14 @@
             </div>
 
             <div class="form-group">
+                <div class="control-label col-md-2">SDT</div>
+                <div class="col-md-10">
+                    @Html.EditorFor(model => model.soDienThoai, new { htmlAttributes = new { @class = "form-control" } })
+                    @Html.ValidationMessageFor(model => model.soDienThoai, "", new { @class = "text-danger" })
+                </div>
+            </div>
+
+            <div class="form-group">
                 <div class="control-label col-md-2">Email</div>
                 <div class="col-md-10">
                     @Html.EditorFor(model => model.email, new { htmlAttributes = new { @class = "form-control" } })
