+++ SHOPHVT/Views/Product/Details.cshtml
@@ -44,20 +44,6 @@
                         @Html.DisplayFor(model => model.moTa)
                     </p>
                     <div class="row mt-4">
-                        <div class="col-md-6">
-                            <div class="form-group d-flex">
-                                <div class="select-wrap">
-                                    <div class="icon"><span class="ion-ios-arrow-down"></span></div>
-                                    <select name="" id="" class="form-control">
-                                        <option value="">Small</option>
-                                        <option value="">Medium</option>
-                                        <option value="">Large</option>
-                                        <option value="">Extra Large</option>
-                                    </select>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="w-100"></div>
                         <div class="input-group col-md-6 d-flex mb-3">
                             <span class="input-group-btn mr-2">
                                 <button type="button" class="quantity-left-minus btn" data-type="minus" data-field="">
@@ -72,9 +58,6 @@
                             </span>
                         </div>
                         <div class="w-100"></div>
-                        <div class="col-md-12">
-                            <p style="color: #000;">600 kg available</p>
-                        </div>
                     </div>
                     <!-- <p><a href="/them-gio-hang?productId=@Model.sanPhamID&quantity=1" class="btn btn-black py-3 px-5">Add to Cart</a></p> -->
                     @using (Html.BeginForm("AddItem", "Cart", FormMethod.Post))
@@ -83,7 +66,7 @@
 
                         <!-- Trường input ẩn để chứa giá trị quantity -->
                         <input type="hidden" id="quantityHidden" name="quantity" class="form-control input-number" value="1" />
-                        <input type="submit" value="Mua hàng" id="bgnGioHang" />
+                        <input type="submit" value="Mua hàng" id="bgnGioHang" class="btn btn-black py-3 px-5" />
                     }
                 </div>
             </div>
