+++ SHOPHVT/Views/Product/Index.cshtml
@@ -34,13 +34,13 @@
             {
                 <div class="col-md-6 col-lg-3 ftco-animate">
                     <div class="product">
-                        <a href="@Url.Action("Details", "TrangChu", new { id = (int)item.SanPhamID })" class="img-prod">
+                        <a href="@Url.Action("Details", "Product", new { id = (int)item.SanPhamID })" class="img-prod">
                             <img class="img-fluid" src="~/images/@item.hinhAnh" alt="Colorlib Template">
                       
                             <div class="overlay"></div>
                         </a>
                         <div class="text py-3 pb-4 px-3 text-center">
-                            <h3><a href="@Url.Action("Details", "TrangChu", new { id = (int)item.SanPhamID })">@item.tenSanPham</a></h3>
+                            <h3 style="height:40px"><a href="@Url.Action("Details", "Product", new { id = (int)item.SanPhamID })">@item.tenSanPham</a></h3>
                             <div class="d-flex">
                                 <div class="pricing">
                                     <p class="price"><span class="price-sale">@item.donGia.ToString("N2") đ</span></p>
@@ -48,9 +48,17 @@
                             </div>
                             <div class="bottom-area d-flex px-3">
                                 <div class="m-auto d-flex">
-                                    <a href="#" class="buy-now d-flex justify-content-center align-items-center mx-1">
+                                    @*<a href="#" class="buy-now d-flex justify-content-center align-items-center mx-1">
                                         <span><i class="ion-ios-cart"></i></span>
-                                    </a>
+                                    </a>*@
+                                    @using (Html.BeginForm("AddItem", "Cart", FormMethod.Post))
+                                    {
+                                        @Html.Hidden("productId", (int)item.SanPhamID)
+                                        <input type="hidden" id="quantityHidden" name="quantity" class="form-control input-number" value="1" />
+                                        <button type="submit" class="buy-now d-flex justify-content-center align-items-center mx-1">
+                                            <span><i class="ion-ios-cart"></i></span>
+                                        </button>
+                                    }
                                 </div>
                             </div>
                         </div>
