+++ SHOPHVT/Views/TrangChu/Dangky.cshtml
@@ -6,7 +6,10 @@
 }
 
 <div class="container body-content">
-    <h2>Đăng ký</h2>
+    <div class="center" style="margin-bottom:20px">
+        <h2>Đăng ký</h2>
+    </div>
+
     @if (TempData["ErrorMessage"] != null)
     {
         <div class="alert alert-danger">
@@ -18,79 +21,81 @@
     {
         @Html.AntiForgeryToken()
 
-        <div class="form-horizontal">
-
+        <div class="row justify-content-center">
 
             @Html.ValidationSummary(true, "", new { @class = "text-danger" })
-            <div class="form-group">
-                <div class="control-label col-md-2">Họ và tên</div>
-                <div class="col-md-10">
+            <div class="col-md-7">
+                <div class="form-group">
+                    <div>Họ và tên</div>
                     @Html.EditorFor(model => model.hoTen, new { htmlAttributes = new { @class = "form-control" } })
                     @Html.ValidationMessageFor(model => model.hoTen, "", new { @class = "text-danger" })
                 </div>
             </div>
 
-            <div class="form-group">
-                <div class="control-label col-md-2">Ngày sinh</div>
-                <div class="col-md-10">
+            <div class="col-md-7">
+                <div class="form-group">
+                    <div>Ngày sinh</div>
                     @Html.EditorFor(model => model.ngaySinh, new { htmlAttributes = new { @class = "form-control" } })
                     @Html.ValidationMessageFor(model => model.ngaySinh, "", new { @class = "text-danger" })
                 </div>
             </div>
 
-            <div class="form-group">
-                <div class="control-label col-md-2">Giới tính</div>
-                <div class="col-md-10">
+            <div class="col-md-7">
+                <div class="form-group">
+                    <div>Giới tính</div>
                     @Html.EditorFor(model => model.gioiTinh, new { htmlAttributes = new { @class = "form-control" } })
                     @Html.ValidationMessageFor(model => model.gioiTinh, "", new { @class = "text-danger" })
                 </div>
             </div>
 
-            <div class="form-group">
-                <div class="control-label col-md-2">Email</div>
-                <div class="col-md-10">
+            <div class="col-md-7">
+                <div class="form-group">
+                    <div>SDT</div>
+                    @Html.EditorFor(model => model.soDienThoai, new { htmlAttributes = new { @class = "form-control" } })
+                    @Html.ValidationMessageFor(model => model.soDienThoai, "", new { @class = "text-danger" })
+                </div>
+            </div>
+
+            <div class="col-md-7">
+                <div class="form-group">
+                    <div>Email</div>
                     @Html.EditorFor(model => model.email, new { htmlAttributes = new { @class = "form-control" } })
                     @Html.ValidationMessageFor(model => model.email, "", new { @class = "text-danger" })
                 </div>
             </div>
 
-            <div class="form-group">
-                <div class="control-label col-md-2">Địa chỉ</div>
-                <div class="col-md-10">
+            <div class="col-md-7">
+                <div class="form-group">
+                    <div>Địa chỉ</div>
                     @Html.EditorFor(model => model.diaChi, new { htmlAttributes = new { @class = "form-control" } })
                     @Html.ValidationMessageFor(model => model.diaChi, "", new { @class = "text-danger" })
                 </div>
             </div>
 
-            <div class="form-group">
-                <div class="control-label col-md-2">Tên đăng nhập</div>
-                <div class="col-md-10">
+            <div class="col-md-7">
+                <div class="form-group">
+                    <div>Tên đăng nhập</div>
                     @Html.EditorFor(model => model.tenDangNhap, new { htmlAttributes = new { @class = "form-control" } })
                     @Html.ValidationMessageFor(model => model.tenDangNhap, "", new { @class = "text-danger" })
                 </div>
             </div>
 
-            <div class="form-group">
-                <div class="control-label col-md-2">Mật khẩu</div>
-                <div class="col-md-10">
+            <div class="col-md-7">
+                <div class="form-group">
+                    <div>Mật khẩu</div>
                     @Html.EditorFor(model => model.matKhau, new { htmlAttributes = new { @class = "form-control" } })
                     @Html.ValidationMessageFor(model => model.matKhau, "", new { @class = "text-danger" })
                 </div>
             </div>
 
-            <div class="form-group">
-                <div class="col-md-offset-2 col-md-10">
+            <div class="col-md-7">
+                <div class="form-group">
                     <input type="submit" value="Đăng ký" class="btn btn-primary" />
-
+                    <a href="@Url.Action("Dangnhap", "TrangChu")" class="px-3">Đăng nhập</a>
                 </div>
             </div>
-
         </div>
     }
-
-    <div>
-        @Html.ActionLink("Back to List", "Index")
-    </div>
 </div>
 @section Scripts {
     @Scripts.Render("~/bundles/jqueryval")
