+++ SHOPHVT/Views/TrangChu/Dangnhap.cshtml
@@ -10,7 +10,6 @@
         <h2>Đăng nhập</h2>
     </div>
 
-
     @if (TempData["ErrorMessage"] != null)
     {
         <div class="alert alert-danger">
@@ -24,52 +23,34 @@
         </div>
     }
 
-
     @using (Html.BeginForm())
     {
         @Html.AntiForgeryToken()
-
-        <div class="form-horizontal justify-content-center ">
-
+        <div class="row justify-content-center">
             @Html.ValidationSummary(true, "", new { @class = "text-danger" })
-
-
-
-
-            <div class="form-group justify-content-center align-items-center">
-                <div class="control-label col-md-2 ">Email</div>
-
-                <div class="col-md-7">
-
+            <div class="col-md-7">
+                <div class="form-group">
+                    <div class="">Email</div>
                     @Html.EditorFor(model => model.email, new { htmlAttributes = new { @class = "form-control" } })
                     @Html.ValidationMessageFor(model => model.email, "", new { @class = "text-danger" })
-
-
                 </div>
             </div>
-
-            <div class="form-group ">
-                <div class="control-label col-md-2">Mật khẩu</div>
-                <div class="col-md-7">
-
+            <div class="col-md-7">
+                <div class="form-group">
+                    <div class="">Mật khẩu</div>
                     @Html.EditorFor(model => model.matKhau, new { htmlAttributes = new { @class = "form-control" } })
-
                     @Html.ValidationMessageFor(model => model.matKhau, "", new { @class = "text-danger" })
                 </div>
             </div>
-
-            <div class="form-group">
-                <div class="col-md-offset-2 col-md-10">
-                    <input type="submit" value="Đăng nhập" class="btn btn-primary" style="width=" 200px" />
+            <div class="col-md-7">
+                <div class="form-group">
+                    <input type="submit" value="Đăng nhập" class="btn btn-primary px-3" width="200"/>
+                    <a href="@Url.Action("Dangky", "TrangChu")" class="px-3">Đăng ký</a>
                 </div>
             </div>
         </div>
         <h3>@ViewBag.Dangnhapfail</h3>
     }
-
-    <div>
-        @Html.ActionLink("Đăng ký", "Dangky")
-    </div>
 </div>
 
 @section Scripts {
